+++ libnetwork/src/main/java/at/aau/ase/libnetwork/androidnetworkwrapper/networking/dto/game_objects/NotePadMessage.java
@@ -4,12 +4,12 @@ import at.aau.ase.libnetwork.androidnetworkwrapper.networking.game.basic_classes
 
 public class NotePadMessage extends BaseMessage{
     private Notepad scores;
-    private short activePlayer;
-    private short betTricks;
+    private int activePlayer;
+    private int betTricks;
 
     public NotePadMessage(){}
 
-    public NotePadMessage(Notepad scores, short activePlayer, short betTricks){
+    public NotePadMessage(Notepad scores, int activePlayer, int betTricks){
         this.scores = scores;
         this.activePlayer = activePlayer;
         this.betTricks = betTricks;
